+++ app/src/main/java/gr/dipae/thesisfitnessapp/domain/user/UserRepository.kt
@@ -43,5 +43,5 @@ interface UserRepository {
     suspend fun getFavoriteSportIds(): List<String>
     suspend fun setFavoriteSportIds(favoritesSports: List<String>)
     suspend fun setUserNewSportChallenges(favoritesSports: List<String>)
-    suspend fun setUserSportChallengeProgress(challengeId: String, progress: Double)
+    suspend fun setUserSportChallengeProgress(challengeId: String, progress: Long)
 }
